Move mocks from src to src/test
