+++ src/test/mock/deep-kernel/Microservice/InstanceMock.js
